First electromagnetic shower cascade is completly possible.
There still have to be added multiple scattering effects and some deflections effects.
The directions and asymetrie of interaction products will be sampled.
There seems to be an issue with the particle cut or a implementation failure on proposal side that if a particle energy is below the cut energy, the particle is simply deleted without taken the particle cut relative difference into account.