Add two tests dealing with 0_mb cross sections in Cascade and ProcessSequence
This commit adds tests to testProcessSequence and testCascade.
In the first test, we create a sequence with only one InteractionProcess with a cross section that is 0.
We check that calling `selectInteraction` does not call the `doInteraction` method of the process.

In the second test, we create a `Cascade` object, based on a sequence including an InteractionProcess with a cross section that is zero.
Again, we check that the `doInteraction` is not called. Furthermore, we check that `Cascade` does not delete the initial particle from the
stack (it should not be deleted since it did not interact!)