Adapt `ProcessSequence` so that if cx_select and cx_sum are zero, we don't select any interaction instead of the first one.
Adapt `Cascade`. In case a discrete interaction has been selected, it will only delete the initial particle from the stack
if `interaction` or `decay` have actually returned that they have called an interaction process or a decay process.
If this is not the case, which means we did not call `doInteract` or `doDecay`, we will keep the particle on the stack.